+++ demo/src/main.tsx
@@ -4,11 +4,12 @@ import { Router, Routes } from 'react-space-router'
 import { Shell } from './Shell'
 import { routes } from './routes'
 import { data } from './data'
+import { routerMode } from './routing'
 import './styles.css'
 
 function App() {
   return (
-    <Router pendingDelayMs={500} data={data} routes={routes}>
+    <Router mode={routerMode} pendingDelayMs={500} data={data} routes={routes}>
       <Shell>
         <Suspense fallback={null}>
           <Routes />